% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{odw}
\alias{odw}
\title{Find the optimal design}
\usage{
odw(
  fixed = ~1,
  random = NULL,
  permute = NULL,
  swap = NULL,
  search = "random",
  maxit = 1,
  start.values = FALSE,
  optimize = NULL,
  group = NULL,
  G.param = NULL,
  R.param = NULL,
  criterion = "A",
  data = NULL
)
}
\arguments{
\item{fixed}{The fixed effects formula.}

\item{random}{The random effects formula.}

\item{permute}{A formula specifying the terms that can be permuted.}

\item{swap}{A formula controlling legal unit exchanges. NOT IMPLEMENTED YET.}

\item{start.values}{A logical value indicating whether to return a list containing
a default variance parameter values. NOT IMPLEMENTED YET.}

\item{optimize}{A character vector identifying the levels of the objective factor
to include in the criterion calculation set. NOT IMPLEMENTED YET.}

\item{group}{A list where each component is a numeric vector specifying the
contiguous fields in \code{data} that are to be considered as a single term. NOT IMPLEMENTED YET.}

\item{G.param, R.param}{A list with pre-specified variance parameter values.}

\item{criterion}{A string specifying the optimality criterion.}

\item{data}{A data frame containing the initial configuration.}
}
\description{
This function mimics the structure given by the \code{odw} package.
}
