% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optrowcoldesign-package.R
\docType{package}
\name{optrowcoldesign-package}
\alias{optrowcoldesign}
\alias{optrowcoldesign-package}
\title{optrowcoldesign: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Emi Tanaka \email{dr.emi.tanaka@gmail.com} (\href{https://orcid.org/0000-0002-1455-259X}{ORCID}) [copyright holder]

}
\keyword{internal}
